% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bound_power.R
\name{bound_power}
\alias{bound_power}
\title{Bound on test's power}
\usage{
bound_power(theta, theta_0, epsilon, n, alpha, power, nsims)
}
\arguments{
\item{theta, theta_0, n}{Characteristics of the public test: it requires a
database of size `n` to achieve `theta` power at a sigificance level of
`theta_0`.}

\item{epsilon}{The privacy parameter.}

\item{alpha}{The significance level for the private test.}

\item{power}{The desired power of the private test.}

\item{nsims}{The number of draws to compute the reference distribution for the
binomial test}
}
\description{
Function that, for given parameters of the public test, will compute an upper
bound on the power of the private test.
}
