% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_hypothesis_test.R
\name{Calculate_Z}
\alias{Calculate_Z}
\title{Privitize binomial data}
\usage{
Calculate_Z(X, epsilon)
}
\arguments{
\item{X}{data from a binomial distribution - typically the number of successes.}

\item{epsilon}{the privacy parameter.}
}
\description{
This function takes in data from a binomial distribution and outputs a
differentally private estimate of that value. Created by Canyon.
}
