% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Canonne_implementation.R
\name{Canonne_power}
\alias{Canonne_power}
\title{Canonne Power}
\usage{
Canonne_power(
  eff,
  n,
  n_zeros,
  d,
  epsilon,
  delta,
  alpha,
  nsims,
  mc.cores = detectCores() - 1
)
}
\arguments{
\item{eff}{Determines the mean of the alternate distribution (which
will be `d - n_zeros` repetitions of `eff`).}

\item{n}{Number of observations (number of rows of the database).}

\item{n_zeros}{Number of dimensions (out of `d`) with no effect.}

\item{d}{Number of dimensions (number of columns of the database).}

\item{epsilon}{The privacy parameter.}

\item{delta}{The privacy parameter.}

\item{alpha}{As in the paper, this is the effect size the test should be able
to detect. THIS IS NOT THE SIGNIFICANCE LEVEL!}

\item{nsims}{The number of times to run the test.}

\item{mc.cores}{Number of cores used for parallelization.}
}
\value{
A double between 0 and 1.
}
\description{
Determines the power of the test proposed by Canonne et al. by simulation.
}
