% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Canonne_implementation.R
\name{Canonne_test}
\alias{Canonne_test}
\title{Canonne Implementation}
\usage{
Canonne_test(data, epsilon, delta, alpha)
}
\arguments{
\item{data}{An `n` by `d` dataframe where `n` is the number of observations
and `d` is the number of dimensions.}

\item{epsilon}{The privacy parameter.}

\item{delta}{The privacy parameter.}

\item{alpha}{As in the paper, this is the effect size the test should be able
to detect. THIS IS NOT THE SIGNIFICANCE LEVEL!}
}
\value{
The result is either to "REJECT" or "FAIL TO REJECT" the null
  hypothesis.
}
\description{
Implementation of Algorithm 3 in Canonne et al. (2019), which tests whether
multivariate Gaussian data comes is centered at 0.
}
