% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Barrientos_implementation.R
\name{cdf_Barr}
\alias{cdf_Barr}
\title{Barreintos reference distribution}
\usage{
cdf_Barr(epsilon, n, a, p, M, reps)
}
\arguments{
\item{epsilon}{The privacy parameter.}

\item{n}{Number of observations in the database (number of rows of the design
matrix).}

\item{a}{Truncation parameter.}

\item{p}{Number of parameters (number of columns of the design matrix).}

\item{M}{Number of subsamples.}

\item{reps}{Number of repetitions.}
}
\value{
a function with outputs on the interval (0,1)
}
\description{
Computes the empirical CDF of the reference distribution for the test
statistic of the test proposed by Barrientos et al.
}
