% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_p_value.R
\name{compute_p_value}
\alias{compute_p_value}
\title{Compute p-value from regression}
\usage{
compute_p_value(df, M, epsilon, beta_number)
}
\arguments{
\item{df}{A dataframe with the response variable as the first column and the
remaining columns explanatory variables.}

\item{M}{The number of subsamples}

\item{epsilon}{The privacy parameter}

\item{beta_number}{The regression coefficient for which to calculate the
p-value. Numering begins at zero (the intercept).}
}
\value{
The output will be a double
}
\description{
Function that computes the p-value for a single regression coefficient.
}
