% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theoretical_power.R
\name{compute_prop_below_threshold_lm}
\alias{compute_prop_below_threshold_lm}
\title{Proportion below threshold for linear regression}
\usage{
compute_prop_below_threshold_lm(num, X, groups, effect_size, theta_0)
}
\arguments{
\item{num}{The index of the subsample of interest.}

\item{X}{A design matrix for a regression. Must have at least two explanatory
variables.}

\item{groups}{A vector of length \code{nrow(X)} with the index of the group
of each row in \code{X}.}

\item{effect_size}{The quotient of the parameter of interest (beta) and the standard
deviation of the noise (sigma).}

\item{theta_0}{The threshold}
}
\value{
The output will be a double between 0 and 1.
}
\description{
Function that computes the proportion of the p-values below a pre-determined
threshold, theta_0, for a given subsample of a design matrix in a regression.
}
