% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theoretical_power.R
\name{compute_prop_below_threshold_normal}
\alias{compute_prop_below_threshold_normal}
\title{Compute proportion below the threshold for the normal test}
\usage{
compute_prop_below_threshold_normal(n, d, effect_size, theta_0)
}
\arguments{
\item{n}{The number of observations (number of rows in the database).}

\item{d}{The number of dimensions (number of columns in the database).}

\item{effect_size}{Determines the mean of the alternate distribution (which
will be `d` repetitions of `effect_size`).}

\item{theta_0}{The threshold}
}
\value{
The output will be a double between 0 and 1.
}
\description{
Function that computes the proportion of the p-values below a pre-determined
threshold, theta_0, for a given subsample of a design matrix in a test of the
mean of multivariate normal data.
}
