% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_hypothesis_test.R
\name{DP_Binom_test_twoside}
\alias{DP_Binom_test_twoside}
\title{Two-sided private binomial hypothesis test.}
\usage{
DP_Binom_test_twoside(n, theta_0, epsilon, Z)
}
\arguments{
\item{n}{number of trials.}

\item{theta_0}{null hypothesis.}

\item{epsilon}{privacy parameter.}

\item{Z}{privatized number of successes.}
}
\value{
outputs the p-value of the hypothesis test.
}
\description{
This function takes in a privatized estimate of a draw from a binomial
distribution and performs a two-sided hypothesis test. Created by Canyon.
}
