% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PB_implementation.R
\name{find_k_alpha0}
\alias{find_k_alpha0}
\title{Function for parameter tuning in Pena-Barrientos}
\usage{
find_k_alpha0(eps, alpha, alpha_min = 0)
}
\arguments{
\item{eps}{The privacy parameter}

\item{alpha}{The significance level}

\item{alpha_min}{The minimum allowable value for alpha0}
}
\description{
The following function is used to find the parameters k, alpha0, and p
for the Pena-Barrientos test
}
