% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_implementation.R
\name{logistic_power}
\alias{logistic_power}
\title{Test of Tests Logistic Power}
\usage{
logistic_power(nsims, n, p, groups, epsilon, effect_size, M, theta_0,
  alpha)
}
\arguments{
\item{nsims}{Number of simulations.}

\item{n}{Number of observations (number of rows of the design matrix)}

\item{p}{Number of parameters (number of columns of the design matrix).
Includes the intercept.}

\item{groups}{A vector with the group number of each observation.}

\item{epsilon}{The privacy parameter.}

\item{effect_size}{The quotient of the parameter of interest (beta) and the
standard deviation of the noise (sigma).}

\item{M}{The number of subsamples when computing p-value.}

\item{theta_0}{The threshold.}

\item{alpha}{The significance level.}
}
\value{
Will return a vector of p-values.
}
\description{
Simulate a p-value computation for logistic regression many times and compute
the power. The design matrix is regenerated for each simulation.
}
