% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theoretical_power.R
\name{optimize_power_ANOVA}
\alias{optimize_power_ANOVA}
\title{Optimize the parameters of our test for ANOVA data}
\usage{
optimize_power_ANOVA(
  n,
  effect_size = 1,
  epsilon = 1,
  alpha = 0.05,
  groups = 3,
  M_max = min(floor(n/groups/2), 50)
)
}
\arguments{
\item{n}{The number of observations (number of rows in the database). Should
be a multiple of three.}

\item{effect_size}{The ratio of the between group variance to the within group
variance. Defaults to 1, as in Couch et al.}

\item{epsilon}{The privacy parameter. Defaults to 1, as in Couch et al.}

\item{alpha}{The significance level. Defaults to 0.05.}

\item{groups}{The number of groups in each subsample for ANOVA.}

\item{M_max}{The maximum M to search over. The optimizer will try all
integers from 1 to M_max}
}
\value{
The output will be a list with the achievable power and the
  corresponding two parameter values
}
\description{
Function that finds the parameters, M and theta_0, of our test that yield
the highest power for a given combination of n, the effect size, and epsilon
}
