% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theoretical_power.R
\name{optimize_power_norm}
\alias{optimize_power_norm}
\title{Optimize the parameters of our test for multivariate normal data}
\usage{
optimize_power_norm(
  effect_size,
  epsilon,
  n,
  alpha = 0.05,
  d = 1,
  n_zeros = 0,
  M_max = min(floor(n/3), 50)
)
}
\arguments{
\item{effect_size}{The quotient of the parameter of interest (mu) and the
standard deviation of the noise (sigma).}

\item{epsilon}{The privacy parameter.}

\item{n}{The number of observations (number of rows in
the database).}

\item{alpha}{The significance level. Defaults to 0.05.}

\item{d}{The number of dimensions (number of columns in
the database). Defaults to 1.}

\item{n_zeros}{The number of entries of the alternative
distribution with mean zero. Defaults to 0.}

\item{M_max}{The maximum M to search over. The optimizer will try all
integers from 1 to M_max}
}
\value{
The output will be a list with the achievable power and the
  corresponding two parameter values
}
\description{
Function that finds the parameters, M and theta_0, of our test that yield
the highest power for a given combination of n, the effect size, and epsilon
}
