% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Barrientos_implementation.R
\name{p_val_Barr}
\alias{p_val_Barr}
\title{Barrientos p-value}
\usage{
p_val_Barr(df, groups, M, a_vector, epsilon, cdfs)
}
\arguments{
\item{df}{A dataframe with the response variable as the first column and the
remaining columns explanatory variables.}

\item{groups}{A vector with the group number of each observation.}

\item{M}{The number of subsamples.}

\item{a_vector}{Vector of values of the truncation parameter.}

\item{epsilon}{The privacy parameter.}

\item{cdfs}{List of the CDFs for each level of truncation. Must be of equal
length to `a_vector`.}
}
\value{
The output will be a vector of length `a_vector`, where each entry
  correpsonds to the p-value from the test with that level of truncation.
}
\description{
Function that computes the p-values from a running of the test proposed by
Barrientos et al. for several choices of the truncation parameter.
}
