% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PB_implementation.R
\name{PB_power_ANOVA}
\alias{PB_power_ANOVA}
\title{Exact power of PB for an ANOVA}
\usage{
PB_power_ANOVA(
  epsilon,
  effect_size,
  n,
  groups = 3,
  alpha = 0.05,
  k = NA,
  alpha0 = NA
)
}
\arguments{
\item{epsilon}{The privacy parameter}

\item{effect_size}{The ratio of the between group variance to the within group
variance.}

\item{n}{The sample size (number of rows)}

\item{groups}{an integer of the number of groups}

\item{alpha}{The significance level}

\item{k, alpha0}{Parameters of PB. If NULL, computed within the function}
}
\description{
The following function gives the exact theoretical power of the Pena-Barrientos
method for a one-way ANOVA.
}
