% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PB_implementation.R
\name{PB_power_norm}
\alias{PB_power_norm}
\title{Exact power of PB for a normal test}
\usage{
PB_power_norm(
  epsilon,
  effect_size,
  n,
  d = 1,
  n_zeros = 0,
  alpha = 0.05,
  k = NA,
  alpha0 = NA
)
}
\arguments{
\item{epsilon}{The privacy parameter}

\item{effect_size}{The quotient of the parameter of interest (mu) and
the standard deviation of the noise (sigma)}

\item{n}{The sample size (number of rows)}

\item{d}{The dimension (number of columns)}

\item{n_zeros}{number of dimensions set to zero}

\item{alpha}{The significance level}

\item{k, alpha0}{Parameters of PB. If NULL, computed within the function}
}
\description{
The following function gives the exact theoretical power of the Pena-Barrientos
method for a normal test.
}
