% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Barrientos_implementation.R
\name{PowerGen_Barr}
\alias{PowerGen_Barr}
\title{Barrientos power}
\usage{
PowerGen_Barr(X, groups, M, a_vector, effect_size, alpha, epsilon, nsims,
  reps)
}
\arguments{
\item{X}{The design matrix.}

\item{groups}{A vector with the group number of each observation.}

\item{M}{The number of subsamples when computing p-value.}

\item{a_vector}{Vector of truncation levels.}

\item{effect_size}{The quotient of the parameter of interest (beta) and the
standard deviation of the noise (sigma).}

\item{alpha}{The significance level.}

\item{epsilon}{The privacy parameter.}

\item{nsims}{Number of simulations for the p-value computation.}

\item{reps}{Number of simulations for the reference distribution computation.}
}
\value{
Will return a vector of p-values.
}
\description{
Simulate a p-value computation for the test proposed by Barrientos et al.
many times and compute the power. This is done for several choices of the
truncation parameter, a. The design matrix is set prior to the simulations.
}
