% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theoretical_power.R
\name{practical_pars_ANOVA}
\alias{practical_pars_ANOVA}
\title{Find practical values of the parameters for an ANOVA}
\usage{
practical_pars_ANOVA(
  effect_size_grid,
  epsilon,
  n,
  rho = 0.8,
  alpha = 0.05,
  groups = 3,
  M_start = min(12, floor(n/2/groups)),
  alpha0_start = 0.2
)
}
\arguments{
\item{effect_size_grid}{Grid of effect sizes to search over}

\item{epsilon}{The privacy parameter.}

\item{n}{The number of observations (number of rows in
the database).}

\item{rho}{The desired power threshold}

\item{alpha}{The significance level. Defaults to 0.05.}

\item{groups}{The number of groups in each subsample for ANOVA.}

\item{M_start}{Starting value of M}

\item{alpha0_start}{Starting value of alpha0}
}
\value{
The output will be a list with the achievable power, the
  corresponding two parameter values, and the minimum effect size
}
\description{
Performs a binary search to find the parameter values that reach power rho at
the smallest effect size. If no effect sizes are sufficient to reach power
rho, returns the parameters for the largest effect size in input grid.
}
