% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theoretical_power.R
\name{practical_pars_norm}
\alias{practical_pars_norm}
\title{Find practical values of the parameters for a normal test}
\usage{
practical_pars_norm(
  effect_size_grid,
  epsilon,
  n,
  rho = 0.8,
  alpha = 0.05,
  d = 1,
  n_zeros = 0,
  M_start = min(12, floor(n/3)),
  alpha0_start = 0.2
)
}
\arguments{
\item{effect_size_grid}{Grid of effect sizes to search over}

\item{epsilon}{The privacy parameter.}

\item{n}{The number of observations (number of rows in
the database).}

\item{rho}{The desired power threshold}

\item{alpha}{The significance level. Defaults to 0.05.}

\item{d}{The number of dimensions (number of columns in
the database). Defaults to 1.}

\item{n_zeros}{The number of entries of the alternative
distribution with mean zero. Defaults to 0.}

\item{M_start}{Starting value of M}

\item{alpha0_start}{Starting value of alpha0}
}
\value{
The output will be a list with the achievable power, the
  corresponding two parameter values, and the minimum effect size
}
\description{
Performs a binary search to find the parameter values that reach power rho at
the smallest effect size. If no effect sizes are sufficient to reach power
rho, returns the parameters for the largest effect size in input grid.
}
