% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theoretical_power.R
\name{practical_power_ANOVA}
\alias{practical_power_ANOVA}
\title{Find the power of our test for an ANOVA in a practical setting}
\usage{
practical_power_ANOVA(
  effect_size,
  epsilon,
  n,
  effect_size_grid = seq(0, 3, 0.2),
  rho = 0.8,
  alpha = 0.05,
  groups = 3,
  M_start = min(12, floor(n/2/groups)),
  alpha0_start = 0.2,
  M_max = floor(n/2/groups)
)
}
\arguments{
\item{effect_size}{The ratio of the between group variance to the within group
variance.}

\item{epsilon}{The privacy parameter.}

\item{n}{The number of observations (number of rows in
the database).}

\item{effect_size_grid}{Grid of effect sizes to search over}

\item{rho}{The desired power threshold}

\item{alpha}{The significance level. Defaults to 0.05.}

\item{groups}{The number of groups in each subsample for ANOVA.}

\item{M_start}{Starting value of M}

\item{alpha0_start}{Starting value of alpha0}

\item{M_max}{The maximum M to search over.}
}
\description{
Function that finds the parameters, M and theta_0, of our test that yield
high power for a given combination of n and epsilon
in a practical setting where the effect size is not known and then computes
the power at a given effect size.
}
