% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_test_power.R
\name{public_power_lm}
\alias{public_power_lm}
\title{Linear regression public power}
\usage{
public_power_lm(X, effect_size, alpha)
}
\arguments{
\item{X}{A design matrix for a regression. Must have at least two columns.}

\item{effect_size}{The quotient of the parameter of interest (beta) and the
standard deviation of the noise (sigma).}

\item{alpha}{The significance level.}
}
\value{
The output will be a double between 0 and 1.
}
\description{
Function that computes the power of public linear regression for a given
design matrix and effect size.
}
