% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_test_power.R
\name{public_power_logistic}
\alias{public_power_logistic}
\title{Logistic regression public power}
\usage{
public_power_logistic(nsims, n, d, effect_size, alpha)
}
\arguments{
\item{nsims}{Number of times to run the test.}

\item{n}{Number of observations (number of rows in the database).}

\item{d}{Number of parameters (number of columns in the database).}

\item{effect_size}{The quotient of the parameter of interest (beta) and the
standard deviation of the noise (sigma).}

\item{alpha}{The significance level.}
}
\value{
The output will be a double between 0 and 1.
}
\description{
Function that computes the power of public logistic regression for a given
effect size.
}
