% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_test_power.R
\name{public_power_normal}
\alias{public_power_normal}
\title{Normal test public power}
\usage{
public_power_normal(n, d, effect_size, alpha)
}
\arguments{
\item{n}{The number of observations (number of rows in the database).}

\item{d}{The number of dimensions (number of columns in the database).}

\item{effect_size}{Determines the mean of the alternate distribution (which
will be `d` repetitions of `effect_size`).}

\item{alpha}{The significance level.}
}
\value{
The output will be a double between 0 and 1.
}
\description{
Function that computes the publuc power for the test of the mean of
multivariate normal data.
}
