% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_hypothesis_test.R
\name{rtulap}
\alias{rtulap}
\title{Random draw from Tulap distribution.}
\usage{
rtulap(n, m, b, q)
}
\arguments{
\item{n}{number of observations.}

\item{m, b, q}{the parameters of the Tulap distribution; described in Awan and
Slavkovic (2018).}
}
\value{
a vector of length \code{n}.
}
\description{
Created by Canyon.
}
