% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_hypothesis_test.R
\name{tulap_cdf}
\alias{tulap_cdf}
\title{CDF of the Tulap distribution.}
\usage{
tulap_cdf(m, b, x)
}
\arguments{
\item{m, b, q}{the parameters of the Tulap distribution; described in Awan and
Slavkovic (2018)}
}
\description{
Created by Canyon.
}
