% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_p_vals.R
\name{simulate_p_vals}
\alias{simulate_p_vals}
\title{Repeatedly simulate a p-value computation.}
\usage{
simulate_p_vals(nsims, X, true_betas, sigma, M, epsilon, beta_number,
  p_value_function = compute_p_value, ...)
}
\arguments{
\item{nsims}{Number of simulations.}

\item{X}{Design matrix from which to compute regression coefficients.}

\item{true_betas}{Vector of true regression parameters. Must be the same
length as the number of columns of \code{X}.}

\item{sigma}{Standard deviation of the error in the regression.}

\item{M}{The number of subsamples when computing p-value.}

\item{epsilon}{The privacy parameter.}

\item{beta_number}{The regression coefficient for which to calculate the
p-value. Numering begins at zero (the intercept).}

\item{p_value_function}{Function to use to compute the p-values.}

\item{...}{Additional options for \code{p_value_function}.}
}
\value{
Will return a vector of p-values.
}
\description{
Simulate a p-value computation many times using choice of computation method
}
