% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theoretical_power.R
\name{theoretical_power}
\alias{theoretical_power}
\title{Compute the theoretical power of our test}
\usage{
theoretical_power(X, groups, M, effect_size, alpha, epsilon, nsims,
  theta_0)
}
\arguments{
\item{X}{A design matrix for a regression. Must have at least two explanatory
variables.}

\item{groups}{A vector of length \code{nrow(X)} with the index of the group
of each row in \code{X}.}

\item{M}{The number of subsamples to partition the data into.}

\item{effect_size}{The quotient of the parameter of interest (beta) and the
standard deviation of the noise (sigma).}

\item{alpha}{The significance level.}

\item{epsilon}{The privacy parameter.}

\item{nsims}{The number of draws from the tulap and binomial with which to
compute the reference distribution.}

\item{theta_0}{The threshold.}
}
\value{
The output will be a double between 0 and 1.
}
\description{
Function that computes the power of our test for a given a design matrix and
a given partitioning into subsamples.
}
