% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theoretical_power.R
\name{theoretical_power}
\alias{theoretical_power}
\title{Theoretical power of our test}
\usage{
theoretical_power(X = NULL, groups = NULL, n = NULL, d = NULL,
  n_zeros = 0, M, effect_size, alpha, epsilon, nsims, theta_0,
  test = "Linear Regression")
}
\arguments{
\item{X}{For regression only. A design matrix with at least two explanatory
variables.}

\item{groups}{For regression only. A vector of length \code{nrow(X)} with the
index of the group of each row in \code{X}.}

\item{n}{For normal test only. The number of observations (number of rows in
the database).}

\item{d}{For normal test only. The number of dimensions (number of columns in
the database).}

\item{n_zeros}{For normal test only. The number of entries of the alternative
distribution with mean zero. Defaults to 0.}

\item{M}{The number of subsamples to partition the data into.}

\item{effect_size}{The quotient of the parameter of interest (beta) and the
standard deviation of the noise (sigma).}

\item{alpha}{The significance level.}

\item{epsilon}{The privacy parameter.}

\item{nsims}{The number of draws from the tulap and binomial with which to
compute the reference distribution.}

\item{theta_0}{The threshold.}

\item{test}{The test to compute the power of. Either "Linear Regression" or
"Normal"}
}
\value{
The output will be a double between 0 and 1.
}
\description{
Function that computes the power of our test for a given a design matrix and
a given partitioning into subsamples.
}
